requirements.txt
anyio==4.11.0
beautifulsoup4==4.14.2
breadability==0.1.20
Brotli==1.1.0
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
ddgs==9.6.1
diskcache==5.6.3
docopt==0.6.2
exceptiongroup==1.3.0
faiss-cpu==1.12.0
filelock==3.20.0
fsspec==2025.9.0
googlesearch-python==1.3.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
llama_cpp_python==0.3.16
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pillow==12.0.0
primp==0.15.0
pycountry==24.6.1
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
sumy==0.11.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typing_extensions==4.15.0
urllib3==2.5.0
